deprecate asdf.asdf
