+++ asdf/_convenience.py
@@ -5,8 +5,8 @@ around _display module code.
 import pathlib
 from contextlib import contextmanager
 
+from ._asdf import AsdfFile, open_asdf
 from ._display import DEFAULT_MAX_COLS, DEFAULT_MAX_ROWS, DEFAULT_SHOW_VALUES, render_tree
-from .asdf import AsdfFile, open_asdf
 
 __all__ = ["info"]
