+++ asdf/_tests/test_deprecated.py
@@ -36,3 +36,10 @@ def test_asdf_util_minversion_deprecation():
 def test_asdf_util_iter_subclasses_deprecation():
     with pytest.warns(AsdfDeprecationWarning, match="asdf.util.iter_subclasses is deprecated"):
         list(asdf.util.iter_subclasses(asdf.AsdfFile))
+
+
+def test_asdf_asdf_deprecation():
+    with pytest.warns(AsdfDeprecationWarning, match="asdf.asdf is deprecated"):
+        if "asdf.asdf" in sys.modules:
+            del sys.modules["asdf.asdf"]
+        import asdf.asdf  # noqa: F401
