+++ asdf/conftest.py
@@ -5,7 +5,7 @@ import pytest
 # We ignore these files because these modules create deprecation warnings on
 # import. When warnings are turned into errors this will completely prevent
 # test collection
-collect_ignore = ["stream.py"]
+collect_ignore = ["asdf.py", "stream.py"]
 
 
 @pytest.fixture(scope="session", autouse=True)
